Move TE stuff out of the blocks package
