+++ src/main/java/vectorwing/farmersdelight/blocks/StoveBlock.java
@@ -25,6 +25,7 @@ import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
+import vectorwing.farmersdelight.tile.StoveTileEntity;
 import vectorwing.farmersdelight.utils.Utils;
 
 import javax.annotation.Nullable;
