+++ src/main/java/vectorwing/farmersdelight/client/gui/CookingPotScreen.java
@@ -8,7 +8,7 @@ import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TextFormatting;
 import vectorwing.farmersdelight.FarmersDelight;
-import vectorwing.farmersdelight.container.CookingPotContainer;
+import vectorwing.farmersdelight.tile.container.CookingPotContainer;
 import vectorwing.farmersdelight.utils.Text;
 
 import java.util.ArrayList;
